ENH: Created ResourceManager class and converted resource inventory file to YAML.
