+++ niceman/interface/base.py
@@ -12,16 +12,11 @@
 
 __docformat__ = 'restructuredtext'
 
-import json
 import sys
 import re
 import textwrap
-import os.path
 
-from ..config import ConfigManager
 from ..ui import ui
-from ..support.exceptions import MissingConfigError, MissingConfigFileError
-from ..resource.base import VALID_RESOURCE_TYPES
 
 from logging import getLogger
 lgr = getLogger('niceman.interface')
@@ -226,179 +221,6 @@ def update_docstring_with_parameters(func, params, prefix=None, suffix=None):
     # assign the amended docs
     func.__doc__ = doc
     return func
-
-
-def get_resource_info(config_path, name, id_=None, type_=None):
-    """
-    Sort through the parameters supplied by the user at the command line and then
-    request the ones that are missing that are needed to find the config and
-    inventory files and then build the config dictionary needed to connect
-    to the environment.
-
-    Parameters
-    ----------
-    config_path : string
-        Path to the niceman.cfg file.
-    name : string
-        Name of the resource
-    id_ : string
-        The identifier of the resource as assigned to it by the backend
-    type_ : string
-        Type of the resource module used to manage the name, e.g.
-        "docker_container".
-
-    Returns
-    -------
-    config : dict
-        The config settings for the name.
-    inventory : dict
-        Inventory of all the managed resources and their configurations.
-    """
-
-    # Get name configuration for this name if it exists
-    # We get the config from inventory first if it exists and then
-    # overlay the default config settings from repronim.cfg
-    cm = get_config_manager(config_path)
-    inventory_path = cm.getpath('general', 'inventory_file')
-    inventory = get_resource_inventory(inventory_path)
-    # XXX: ATM mixes creation with querying existing resources.
-    #      IMHO (yoh) should just query, and leave creation to a dedicated function
-    # TODO: query could be done via ID
-    # TODO: check that if both name and id provided -- they are as it is in
-    #       inventory
-    # TODO:  if no name or id provided, then fail since this function
-    #        is not created to return a list of resources for a given type ATM
-    if name in inventory:
-        # XXX so what is our convention here on SMTH-SMTH defining the type?
-        config = dict(cm.items(inventory[name]['type'].split('-')[0]))
-        config.update(inventory[name])
-    elif type_ and type_ in VALID_RESOURCE_TYPES:
-        config = dict(cm.items(type_.split('-')[0]))
-    else:
-        type_ = ui.question(
-            "Enter a resource type",
-            # TODO: decision on type of a container, if needed
-            # needs to be done outside, and should be configurable
-            # or follow some heuristic (e.g. if name starts with a
-            # known type, e.g. docker-
-            default="docker-container"
-        )
-        config = {}
-        if type_ not in VALID_RESOURCE_TYPES:
-            raise MissingConfigError(
-                "Resource type '{}' is not valid".format(type_))
-
-    # Overwrite config settings with those from the command line.
-    config['name'] = name
-    if type_:
-        config['type'] = type_
-    if id_:
-        config['id'] = id_
-
-    return config, inventory
-
-
-def get_config_manager(config_path=None):
-    """
-    Returns the information stored in the niceman.cfg file.
-
-    Parameters
-    ----------
-    config_path : string
-        Path to the niceman.cfg file. (optional)
-
-    Returns
-    -------
-    cm : ConfigManager object
-        Information stored in the niceman.cfg file.
-    """
-    def get_cm(config_path):
-        if config_path:
-            cm = ConfigManager([config_path], False)
-        else:
-            cm = ConfigManager()
-        return cm
-
-    # Look for a niceman.cfg file in the local directory if none given.
-    if not config_path and os.path.isfile('niceman.cfg'):
-        config_path = 'niceman.cfg'
-    cm = get_cm(config_path=config_path)
-    if not config_path and len(cm._sections) == 1:
-        config = ui.question("Enter a config file", default="niceman.cfg")
-        cm = get_cm(config_path=config)
-    if len(cm._sections) == 1:
-        raise MissingConfigFileError(
-            "Unable to locate config file: {}".format(config_path))
-
-    return cm
-
-
-def get_resource_inventory(inventory_path):
-    """
-    Returns a dictionary containing the config information for all resources
-    created by niceman.
-
-    Parameters
-    ----------
-    inventory_path : string
-        Path to the inventory file which is declared in the niceman.cfg file.
-
-    Returns
-    -------
-    inventory : dict
-        Hash whose key is resource name and value is the config settings for
-        the resource.
-    """
-    if not inventory_path:
-        raise MissingConfigError(
-            "No resource inventory file declared in niceman.cfg")
-
-    # Create inventory file if it does not exist.
-    if not os.path.isfile(inventory_path):
-        lgr.info("Creating resources inventory file %s", inventory_path)
-        # initiate empty inventory
-        set_resource_inventory({'_path': inventory_path})
-
-    with open(inventory_path, 'r') as fp:
-        inventory = json.load(fp)
-
-    inventory['_path'] = inventory_path
-    return inventory
-
-
-def set_resource_inventory(inventory):
-    """
-    Save the resource inventory to a file. The location of the file is
-    declared in the niceman.cfg file.
-
-    Parameters
-    ----------
-    inventory : dict
-        Hash whose key is the name of the resource and value is the config
-        settings of the resource.
-    """
-
-    # Operate on a copy so there is no side-effect of modifying original
-    # inventory
-    inventory = inventory.copy()
-    inventory_path = inventory.pop('_path')
-
-    for key in list(inventory):  # go through a copy of all keys since we modify
-
-        # A resource without an ID has been deleted.
-        inventory_item = inventory[key]
-        if 'id' in inventory_item and not inventory_item['id']:
-            del inventory[key]
-
-        # Remove AWS credentials
-        # XXX(yoh) where do we get them from later?
-        for secret_key in ('access_key_id', 'secret_access_key'):
-            if secret_key in inventory_item:
-                del inventory_item[secret_key]
-
-    with open(inventory_path, 'w') as fp:
-        json.dump(inventory, fp, indent=2, sort_keys=True)
-
 
 class Interface(object):
     """Base class for interface implementations"""
