+++ niceman/interface/tests/test_create.py
@@ -22,8 +22,8 @@ def test_create_interface(niceman_cfg_path):
     """
 
     with patch('docker.Client') as client, \
-        patch('niceman.interface.base.set_resource_inventory'), \
-        patch('niceman.interface.base.get_resource_inventory') as get_inventory, \
+        patch('niceman.resource.ResourceManager.set_inventory'), \
+        patch('niceman.resource.ResourceManager.get_inventory') as get_inventory, \
         swallow_logs(new_level=logging.DEBUG) as log:
 
         client.return_value = MagicMock(
