+++ niceman/tests/files/niceman.cfg
@@ -1,5 +1,5 @@
 [general]
-inventory_file = ~/.config/niceman/inventory.json
+inventory_file = ~/.config/niceman/inventory.yml
 
 [aws]
 access_key_id = AWS-ACCESS-KEY
